rename blackbox, document
