+++ engine/requirements.txt
@@ -1,0 +1,6 @@
+joblib==1.2.0
+pandas==1.5.3
+libpysal==4.7.0
+xarray==2023.1.0
+pyarrow==11.0.0
+scikit-learn==1.2.0
