+++ python/Readme.md
@@ -1,0 +1,7 @@
+# Research code
+
+This folder contains research code used to create the predictive engine. It is more of
+archive of individual notebooks rather than a workflow but all code that has been used
+is here. Please refer to the individual notebooks for more details on their contents.
+
+Some notebooks contain only prototypes and the final code may live elsewhere.
